pwdb - Introduce --import-file and --export-file
Initial support for importing and exporting the entire database from and
to an encrypted JSON file in which records are **not** indivicually encrypted,
but only protected by the file overrall encryption.

The implementation is messy, with most logic having been dumped directly into
main(). Interplay with other options is imperfect and may lead to behaviour
which is unexpected for some option combinations. The new features seem to
work, but a fair bit of cleanup and re-organization should happen prior to
merging.