gpgh - Introduce struct sig_verify_result and op_verify_result()
Handling signature verification results is a pain because gpgme gives
a list of complex difficult to copy objects whos lifetimes expire at
the next operation on the context. Also, the only **reliable** way to get
the uid of the key corresponding to a result is to look it up by uid,
which takes an operation on the context, which invalidates the list.

Introduce a struct to which the important parts ot the result object may
be copied and a helper function to build a list of these structs from the
result object so that the results may be used freely without invalidation.

Simplify pwdb-main check_gpg_verify_result() using these new facilities.

Also simplify keylist to be just a list of unique_ptr's, getting rid of the
error-prone shared_ptr wrapper. Changed copies and copy assignments to moves
and move assignments as necessary.